Rename file to match ClassName, add documentation and fix linting warnings
