--- src/include/utility/Utilities.hpp
@@ -1,21 +1,0 @@
-/**
- * \file Utilities.hpp
- * \author
- * \brief
- * \details
- *
- * \copyright See the LICENSE file.
- */
-
-#include <string>
-
-namespace utilities {
-class Utils {
-public:
-    static void setupEasyLogging(const std::string &configFile);
-    static bool checkIfFileExists(const std::string &fileName);
-    static bool checkFileEnding(const std::string &fileName);
-    static bool
-    askToContinue(const std::string &prompt = "Do you want to continue? (Y/N)\n");
-};
-} // namespace utilities
